+++ src/wildboar/distance/_elastic.pyx
@@ -6,9 +6,11 @@
 # Authors: Isak Samsten
 # License: BSD 3 clause
 
+import warnings
+
 import numpy as np
 
-from libc.math cimport INFINITY, exp, fabs, floor, sqrt
+from libc.math cimport INFINITY, NAN, exp, fabs, floor, isnan, sqrt
 from libc.stdlib cimport free, labs, malloc
 from libc.string cimport memcpy
 
@@ -929,7 +931,7 @@ cdef double lcss_distance(
     const double *Y,
     Py_ssize_t y_length,
     Py_ssize_t r,
-    double threshold,
+    double epsilon,
     double *cost,
     double *cost_prev,
     double *weight_vector,
@@ -967,7 +969,7 @@ cdef double lcss_distance(
             if weight_vector != NULL:
                 w = weight_vector[labs(i - j)]
 
-            if v <= threshold:
+            if v <= epsilon:
                 cost[j] = w + y
             else:
                 cost[j] = max(z, x)
@@ -978,6 +980,90 @@ cdef double lcss_distance(
         cost, cost_prev = cost_prev, cost
 
     return 1 - (cost_prev[y_length - 1] / min(x_length, y_length))
+
+
+cdef double lcss_subsequence_distance(
+    const double *S,
+    Py_ssize_t s_length,
+    const double *T,
+    Py_ssize_t t_length,
+    Py_ssize_t r,
+    double epsilon,
+    double *cost,
+    double *cost_prev,
+    double *weight_vector,
+    Py_ssize_t *index,
+) nogil:
+    cdef double dist = 0
+    cdef double min_dist = INFINITY
+
+    cdef Py_ssize_t i
+    for i in range(t_length - s_length + 1):
+        dist = lcss_distance(
+            S,
+            s_length,
+            T + i,
+            s_length,
+            r, 
+            epsilon,
+            cost, 
+            cost_prev, 
+            weight_vector,
+        )
+
+        if dist < min_dist:
+            if index != NULL:
+                index[0] = i
+            min_dist = dist
+
+    return min_dist
+
+
+cdef Py_ssize_t lcss_subsequence_matches(
+    const double *S,
+    Py_ssize_t s_length,
+    const double *T,
+    Py_ssize_t t_length,
+    Py_ssize_t r,
+    double epsilon,
+    double *cost,
+    double *cost_prev,
+    double *weight_vector,
+    double threshold,
+    double **distances,
+    Py_ssize_t **matches
+) nogil:
+    cdef double dist = 0
+    cdef Py_ssize_t capacity = 1
+    cdef Py_ssize_t tmp_capacity
+    cdef Py_ssize_t i
+    cdef Py_ssize_t n_matches = 0
+
+    matches[0] = <Py_ssize_t*> malloc(sizeof(Py_ssize_t) * capacity)
+    distances[0] = <double*> malloc(sizeof(double) * capacity)
+
+    for i in range(t_length - s_length + 1):
+        dist = lcss_distance(
+            S,
+            s_length,
+            T + i,
+            s_length,
+            r, 
+            epsilon,
+            cost, 
+            cost_prev, 
+            weight_vector,
+        )
+
+        if dist <= threshold:
+            tmp_capacity = capacity
+            realloc_array(<void**> matches, n_matches, sizeof(Py_ssize_t), &tmp_capacity)
+            realloc_array(<void**> distances, n_matches, sizeof(double), &capacity)
+            matches[0][n_matches] = i
+            distances[0][n_matches] = dist
+            n_matches += 1
+
+    return n_matches
 
 
 cdef double erp_distance(
@@ -1251,14 +1337,15 @@ cdef double twe_distance(
     return cost_prev[y_length - 1]
 
 
-
-
-
 cdef Py_ssize_t _compute_warp_width(Py_ssize_t length, double r) nogil:
     if r == 1:
         return length - 1
     else:
         return <Py_ssize_t> floor(length * r)
+
+
+cdef inline Py_ssize_t _compute_r(Py_ssize_t length, double r) nogil:
+    return <Py_ssize_t> max(floor(length * r), 1)
 
 
 cdef class ScaledDtwSubsequenceDistanceMeasure(ScaledSubsequenceDistanceMeasure):
@@ -1745,6 +1832,93 @@ cdef class DtwSubsequenceDistanceMeasure(SubsequenceDistanceMeasure):
         )
 
 
+cdef class LcssSubsequenceDistanceMeasure(SubsequenceDistanceMeasure):
+
+    cdef double *cost
+    cdef double *cost_prev
+    cdef double r
+    cdef double epsilon
+    
+    def __init__(self, double r=1.0, double epsilon=1.0):
+        check_scalar(r, "r", float, min_val=0.0, max_val=1.0)
+        check_scalar(epsilon, "epsilon", float, min_val=0)
+        self.r = r
+        self.epsilon = epsilon
+
+    def __cinit__(self, *args, **kwargs):
+        self.cost = NULL
+        self.cost_prev = NULL
+
+    def __reduce__(self):
+        return self.__class__, (self.r, self.epsilon)
+
+    def __dealloc__(self):
+        self._free()
+
+    cdef void _free(self) nogil:
+        if self.cost != NULL:
+            free(self.cost)
+            self.cost = NULL
+
+        if self.cost_prev != NULL:
+            free(self.cost_prev)
+            self.cost_prev = NULL
+
+    cdef int reset(self, TSArray X) nogil:
+        self._free()
+        self.cost = <double*> malloc(sizeof(double) * X.shape[2])
+        self.cost_prev = <double*> malloc(sizeof(double) * X.shape[2])
+        if self.cost == NULL or self.cost_prev == NULL:
+            return -1
+
+
+    cdef double _distance(
+        self,
+        const double *s,
+        Py_ssize_t s_len,
+        const double *x,
+        Py_ssize_t x_len,
+        Py_ssize_t *return_index=NULL,
+    ) nogil:
+        return lcss_subsequence_distance(
+            s, 
+            s_len,
+            x,
+            x_len,
+            _compute_r(s_len, self.r),
+            self.epsilon,
+            self.cost,
+            self.cost_prev,
+            NULL,
+            return_index,
+        )
+
+    cdef Py_ssize_t _matches(
+        self,
+        double *s,
+        Py_ssize_t s_len,
+        double *x,
+        Py_ssize_t x_len,
+        double threshold,
+        double **distances,
+        Py_ssize_t **indicies,
+    ) nogil:
+        return lcss_subsequence_matches(
+            s,
+            s_len,
+            x,
+            x_len,
+            _compute_r(s_len, self.r),
+            self.epsilon,
+            self.cost,
+            self.cost_prev,
+            NULL,
+            threshold,
+            distances,
+            indicies,
+        )
+
+
 cdef class DtwDistanceMeasure(DistanceMeasure):
 
     cdef double *cost
@@ -1980,18 +2154,27 @@ cdef class LcssDistanceMeasure(DistanceMeasure):
     cdef double *cost_prev
     cdef Py_ssize_t warp_width
     cdef double r
-    cdef double threshold
+    cdef double epsilon
     
-    def __cinit__(self, double r=1.0, double threshold=1.0, *args, **kwargs):
+    def __init__(self, double r=1.0, double epsilon=1.0, threshold=NAN):
+        # TODO(1.4): remove deprecated
+        if not isnan(threshold):
+            warnings.warn(
+                "The parameter threshold has been renamed to epsilon in 1.2 and will be "
+                "removed in 1.4.",
+                DeprecationWarning
+            )
+            epsilon = threshold
+
         check_scalar(r, "r", float, min_val=0.0, max_val=1.0)
-        check_scalar(threshold, "threshold", float, min_val=0)
+        check_scalar(epsilon, "epsilon", float, min_val=0)
         self.r = r
-        self.threshold = threshold# * threshold
+        self.epsilon = epsilon
         self.cost = NULL
         self.cost_prev = NULL
 
     def __reduce__(self):
-        return self.__class__, (self.r, self.threshold)
+        return self.__class__, (self.r, self.epsilon)
 
     def __dealloc__(self):
         self.__free()
@@ -2025,7 +2208,7 @@ cdef class LcssDistanceMeasure(DistanceMeasure):
             y,
             y_len,
             self.warp_width,
-            self.threshold,
+            self.epsilon,
             self.cost,
             self.cost_prev,
             NULL,
@@ -2043,12 +2226,12 @@ cdef class WeightedLcssDistanceMeasure(LcssDistanceMeasure):
     cdef double g
     cdef double *weights
 
-    def __cinit__(self, double r=1.0, threshold=1.0, double g=0.05):
+    def __init__(self, double r=1.0, epsilon=1.0, double g=0.05, threshold=NAN):
         check_scalar(r, "r", float, min_val=0.0, max_val=1.0)
-        check_scalar(threshold, "threshold", float, min_val=0.0)
+        check_scalar(epsilon, "epsilon", float, min_val=0.0)
         self.weights = NULL
         self.g = g
-        self.threshold = threshold
+        self.epsilon = epsilon
 
     def __reduce__(self):
         return self.__class__, (self.r, self.g)
@@ -2081,7 +2264,7 @@ cdef class WeightedLcssDistanceMeasure(LcssDistanceMeasure):
             y,
             y_len,
             self.warp_width,
-            self.threshold,
+            self.epsilon,
             self.cost,
             self.cost_prev,
             self.weights,